+++ java/vector/src/main/java/org/apache/arrow/vector/BitVector.java
@@ -249,7 +249,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param index position of element
    * @return element at given index
    */
-  public int get(int index) throws IllegalStateException {
+  public int get(long index) throws IllegalStateException {
     if (NULL_CHECKING_ENABLED && isSet(index) == 0) {
       throw new IllegalStateException("Value at index is null");
     }
@@ -263,7 +263,7 @@ public class BitVector extends BaseFixedWidthVector {
    *
    * @param index position of element
    */
-  public void get(int index, NullableBitHolder holder) {
+  public void get(long index, NullableBitHolder holder) {
     if (isSet(index) == 0) {
       holder.isSet = 0;
       return;
@@ -318,7 +318,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param index position of element
    * @param value value of element
    */
-  public void set(int index, int value) {
+  public void set(long index, int value) {
     BitVectorHelper.setValidityBitToOne(validityBuffer, index);
     if (value != 0) {
       BitVectorHelper.setValidityBitToOne(valueBuffer, index);
@@ -335,7 +335,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param index  position of element
    * @param holder nullable data holder for value of element
    */
-  public void set(int index, NullableBitHolder holder) throws IllegalArgumentException {
+  public void set(long index, NullableBitHolder holder) throws IllegalArgumentException {
     if (holder.isSet < 0) {
       throw new IllegalArgumentException();
     } else if (holder.isSet > 0) {
@@ -356,7 +356,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param index  position of element
    * @param holder data holder for value of element
    */
-  public void set(int index, BitHolder holder) {
+  public void set(long index, BitHolder holder) {
     BitVectorHelper.setValidityBitToOne(validityBuffer, index);
     if (holder.value != 0) {
       BitVectorHelper.setValidityBitToOne(valueBuffer, index);
@@ -366,40 +366,40 @@ public class BitVector extends BaseFixedWidthVector {
   }
 
   /**
-   * Same as {@link #set(int, int)} except that it handles the
+   * Same as {@link #set(long, int)} except that it handles the
    * case when index is greater than or equal to existing
    * value capacity {@link #getValueCapacity()}.
    *
    * @param index position of element
    * @param value value of element
    */
-  public void setSafe(int index, int value) {
+  public void setSafe(long index, int value) {
     handleSafe(index);
     set(index, value);
   }
 
   /**
-   * Same as {@link #set(int, NullableBitHolder)} except that it handles the
+   * Same as {@link #set(long, NullableBitHolder)} except that it handles the
    * case when index is greater than or equal to existing
    * value capacity {@link #getValueCapacity()}.
    *
    * @param index  position of element
    * @param holder nullable data holder for value of element
    */
-  public void setSafe(int index, NullableBitHolder holder) throws IllegalArgumentException {
+  public void setSafe(long index, NullableBitHolder holder) throws IllegalArgumentException {
     handleSafe(index);
     set(index, holder);
   }
 
   /**
-   * Same as {@link #set(int, BitHolder)} except that it handles the
+   * Same as {@link #set(long, BitHolder)} except that it handles the
    * case when index is greater than or equal to existing
    * value capacity {@link #getValueCapacity()}.
    *
    * @param index  position of element
    * @param holder data holder for value of element
    */
-  public void setSafe(int index, BitHolder holder) {
+  public void setSafe(long index, BitHolder holder) {
     handleSafe(index);
     set(index, holder);
   }
@@ -409,7 +409,7 @@ public class BitVector extends BaseFixedWidthVector {
    *
    * @param index position of element
    */
-  public void setNull(int index) {
+  public void setNull(long index) {
     handleSafe(index);
     // not really needed to set the bit to 0 as long as
     // the buffer always starts from 0.
@@ -424,7 +424,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param isSet 0 for NULL value, 1 otherwise
    * @param value element value
    */
-  public void set(int index, int isSet, int value) {
+  public void set(long index, int isSet, int value) {
     if (isSet > 0) {
       set(index, value);
     } else {
@@ -433,7 +433,7 @@ public class BitVector extends BaseFixedWidthVector {
   }
 
   /**
-   * Same as {@link #set(int, int, int)} except that it handles the case
+   * Same as {@link #set(long, int, int)} except that it handles the case
    * when index is greater than or equal to current value capacity of the
    * vector.
    *
@@ -441,7 +441,7 @@ public class BitVector extends BaseFixedWidthVector {
    * @param isSet 0 for NULL value, 1 otherwise
    * @param value element value
    */
-  public void setSafe(int index, int isSet, int value) {
+  public void setSafe(long index, int isSet, int value) {
     handleSafe(index);
     set(index, isSet, value);
   }
@@ -451,7 +451,7 @@ public class BitVector extends BaseFixedWidthVector {
    *
    * @param index position of element
    */
-  public void setToOne(int index) {
+  public void setToOne(long index) {
     BitVectorHelper.setValidityBitToOne(validityBuffer, index);
     BitVectorHelper.setValidityBitToOne(valueBuffer, index);
   }
@@ -462,7 +462,7 @@ public class BitVector extends BaseFixedWidthVector {
    *
    * @param index position of the element
    */
-  public void setSafeToOne(int index) {
+  public void setSafeToOne(long index) {
     handleSafe(index);
     setToOne(index);
   }
